+++ online-docs/pages/whats-new.rst
@@ -3,6 +3,12 @@ What's new
 
 Following is a brief list of important updates to the COMPAS code.  A complete record of changes can be found in the file ``changelog.h``.
 
+**03.14.00 Mar 3, 2025**
+
+* Updates to improve convergence without sacrificing computational speed
+* Capped total wind mass loss rate at MAXIMUM_WIND_MASS_LOSS_RATE (set to 0.1 Msol/yr) for all prescriptions
+* Changed order of calls in SSE evolution to better match BSE evolution
+
 **03.13.02 Feb 19, 2025**
 
 * Replaced name of ``SHIKAUCHI`` main sequence core mass prescription with ``BRCEK``.